Remove dependency on void crate
Just writing out the functionality that the void crate provides actually
takes less lines than taking it as a dependency. The void crate provides
more, but it is not used by this crate, so it is only wasteful to
compile.

Version 1.0.2 of the void crate takes 2.4 KiB on my disk. That is not a
lot in terms of disk space, nor in terms of bandwith -- for me. But now
that this crate is a transitive dependency of a popular crate, it is
being downloaded more than 3000 times per day according to crates.io,
wasting 0.2 GiB per month. Still not a lot, but not necessary either.
Furthermore, the extra dependency adds an extra network roundtrip to
a clean build with no cached dependencies (such as many CI builds).
On a bad connection that can make a noticeable difference.

Finally, one of the advantages of relying on an external crate -- being
able to independently and centrally fix bugs -- does not apply so much
here. Those few lines of code are more likely in the "obviously no bugs"
category than in the "no obvious bugs" category.